Adding CharacterOverview view logic & some clean up.
