+++ src/components/character-overview/character-overview.module.css
@@ -1,0 +1,19 @@
+
+.characterOverview {
+    position: absolute;
+    top: 0px;
+    left: 0px;
+    width: 100%;
+    height: 100%;
+    background-color: rgba( 0, 0, 0, 0.9 );
+    z-index: 2;
+    transition: visibility 0s 300ms, opacity 300ms ease;
+    opacity: 0;
+    visibility: hidden;
+}
+
+.characterOverview.open {
+    opacity: 1;
+    visibility: visible;
+    transition: visibility 0s 0ms, opacity 300ms ease;
+}
