+++ src/components/character-overview/index.tsx
@@ -1,0 +1,3 @@
+
+import { CharacterOverview } from './CharacterOverview';
+export { CharacterOverview };
