+++ src/components/crosshair/Corsshair.jsx
@@ -1,0 +1,14 @@
+
+import React from 'react';
+
+import styles from './crosshair.module.css';
+
+//
+
+export const Crosshair = () => (
+
+    <div className={ styles.crosshair } id="crosshair">
+        <img src="./assets/crosshair.svg" width={ 30 } height={ 30 } />
+    </div>
+
+);
