+++ src/components/crosshair/index.jsx
@@ -1,0 +1,3 @@
+
+import { Crosshair } from './Corsshair';
+export { Crosshair };
