+++ src/components/diorama/Diorama.tsx
@@ -1,0 +1,13 @@
+
+import React from 'react';
+
+import styles from './diorama.module.css';
+//
+
+export const Diorama = () => {
+
+    return (
+        <div className={ styles.diorama } />
+    );
+
+};
