+++ src/components/diorama/diorama.module.css
@@ -1,0 +1,4 @@
+
+.diorama {
+
+}
