+++ src/components/diorama/index.tsx
@@ -1,0 +1,3 @@
+
+import { Diorama } from './Diorama';
+export { Diorama };
