+++ src/components/inventory/Inventory.tsx
@@ -5,13 +5,24 @@ import styles from './inventory.module.css';
 
 //
 
-export const Inventory = () => {
+export const Inventory = ({ openCharacterOverview }) => {
+
+    const handleInventoryBtn = ( event ) => {
+
+        event.stopPropagation();
+        openCharacterOverview( true );
+
+    };
+
+    //
 
     return (
         <div className={ styles.inventory } >
 
+            <div className={ styles.inventoryBtn } onClick={ handleInventoryBtn } >Inventory ['i' btn]</div>
+
             <div className={ styles.weapon } >
-                <div className={ styles.weponStats } >99</div>
+                <div className={ styles.weaponStats } >99</div>
             </div>
 
         </div>
