+++ src/components/inventory/inventory.module.css
@@ -4,7 +4,7 @@
     bottom: 0px;
     right: 0px;
     width: 300px;
-    height: 200px;
+    height: 280px;
     z-index: 1;
 }
 
@@ -19,13 +19,13 @@
     border-radius: 1px solid #000;
 }
 
-.weponStats {
+.weaponStats {
     position: absolute;
     left: 5px;
     bottom: 5px;
     width: 40px;
     height: 40px;
-    border-radius: 50%;
+    border-radius: 100%;
     background-color: rgba( 255, 0, 0, 0.5 );
     border: 2px solid #000;
     text-align: center;
@@ -33,4 +33,14 @@
     font-weight: bold;
     font-size: 20px;
     line-height: 38px;
+}
+
+.inventoryBtn {
+    position: absolute;
+    top: 0px;
+    right: 0px;
+    width: 100px;
+    height: 50px;
+    background-color: rgba( 0, 0, 0, 0.5 );
+    cursor: pointer;
 }
