+++ src/components/magic-menu/index.jsx
@@ -1,0 +1,3 @@
+
+import { MagicMenu } from './MagicMenu';
+export { MagicMenu };
