+++ src/components/magic-menu/magic-menu.module.css
@@ -1,0 +1,65 @@
+.MagicMenu {
+    position: absolute;
+    top: 0;
+    bottom: 0;
+    left: 0;
+    right: 0;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    z-index: 1;
+    visibility: hidden;
+}
+
+.MagicMenu .container {
+    display: flex;
+    flex-direction: column;
+}
+
+.MagicMenu .container .textarea {
+    height: auto;
+    margin-bottom: 10px;
+    padding: 20px;
+    background-color: #EEE;
+    border: 0;
+    font-size: 16px;
+    outline: none;
+    box-shadow: 0 5px 20px #00000080;
+}
+
+.MagicMenu .container .output {
+    height: 500px;
+    margin-bottom: 10px;
+    padding: 20px;
+    background-color: #333;
+    border: 0;
+    color: #FFF;
+    font-size: 16px;
+    outline: none;
+}
+
+.MagicMenu .container .buttons {
+    display: flex;
+}
+.MagicMenu .container .button {
+    padding: 15px 20px;
+    border: 0;
+    background-color: #7E57C2;
+    color: #FFF;
+    font-size: 16px;
+}
+.MagicMenu .container .button.disabled {
+    background-color: #CCC;
+}
+.MagicMenu .container .button + .button {
+    margin-left: 5px;
+}
+
+.MagicMenu.open {
+    visibility: visible;
+}
+
+.button.loading {
+    background-color: #CCC;
+}
