+++ src/components/player-zone/PlayerZone.jsx
@@ -1,4 +1,5 @@
 
+import classNames from 'classnames';
 import React from 'react';
 
 import styles from './player-zone.module.css';
@@ -11,8 +12,12 @@ export const PlayerZone = () => {
         <div className={ styles.playerZone } >
 
             <div className={ styles.avatar } />
-            <div className={ styles.healthBar } >
-                <div className={ styles.healthBarFill } />
+            <div className={ classNames( styles.progressBar, styles.manaBar ) } >
+                <div className={ styles.progressBarFill } />
+            </div>
+
+            <div className={ classNames( styles.progressBar, styles.healthBar ) } >
+                <div className={ styles.progressBarFill } />
             </div>
 
         </div>
